test: refactor a11y tests
It seems since the move to Xcode 26.0 and the use of iPhone 17 Pro as simulator the UI tests made with interactions on components are broken.
The state retrieved after a tap is not the same, or the element is not tapped.
Thus, instead of making a tpa on the component, we use the chips for state changes. This solution is a bit worse because we do not make interaction on component itself, but the a11y values are still tested

Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>