ci: remove `periphery` GitHub Actions workflow
Periphery is a nice tool but has some issues with numerous false positive cases.
It needs also to compile projects for all targets.
In GitHub Actions it failed sometimes to build for other OS than iOS and iPadOS.
In addition it consumes time and budget, and because of false positive cases errors
are accepted and lost in noise with warnings.
This Periphery must be used locally to have a manuel check of the results while false
positive cases remain.

Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>